refactor: move cli processes to cli.go
for testing
